♻️ Module "kotlin-codegen" now contains all the code-generation stuff
