+++ kotlin-codegen/src/main/graphql/launches/schema.sdl
