+++ kotlin-codegen/src/main/kotlin/playground/KotlinxSerialization.kt
@@ -2,11 +2,13 @@
 
 package playground.kotlinx.serialization
 
-import kotlinx.serialization.*
-import kotlinx.serialization.json.*
+import kotlinx.serialization.ExperimentalSerializationApi
+import kotlinx.serialization.Serializable
+import kotlinx.serialization.decodeFromString
+import kotlinx.serialization.encodeToString
+import kotlinx.serialization.json.Json
 import kotlinx.serialization.properties.Properties
 import kotlinx.serialization.properties.encodeToMap
-import kotlinx.serialization.Serializable
 import playground.shouldBe
 
 /**
