+++ kotlin-jvm/build.gradle.kts
@@ -1,27 +1,16 @@
-import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
-
 plugins {
     kotlin("jvm")
-    kotlin("kapt")
-    kotlin("plugin.serialization")
-    id("com.squareup.sqldelight")
-    id("com.apollographql.apollo")
+    // NOTE: IF the library does code generation, either with kapt or a gradle/compiler plugin
+    // NOTE: THEN it belongs in "kotlin-codegen", not here!
 }
 
 tasks.withType<JavaExec> {
     classpath = sourceSets["main"].runtimeClasspath
 }
 
-sqldelight {
-    database("AppDatabase") {
-        packageName = "util"
-    }
-    linkSqlite = false
-}
-
 // File build.gradle.kts
 dependencies {
-
+    implementation(project(":kotlin-codegen"))
     // Keep dependencies sorted to minimize merge conflicts on pull-requests!
     implementation("com.beust:klaxon:_")
     implementation ("com.github.ajalt:mordant:_")
@@ -29,11 +18,8 @@ dependencies {
     implementation("com.github.ajalt.clikt:clikt:_")
     implementation("com.h2database:h2:_")
     implementation("com.h2database:h2:_")
-    implementation("com.squareup.moshi:moshi:_")
     implementation("com.squareup.sqldelight:sqlite-driver:_")
     implementation("io.github.serpro69:kotlin-faker:_")
-    implementation("com.apollographql.apollo:apollo-runtime:_")
-    implementation("com.apollographql.apollo:apollo-coroutines-support:_")
     implementation("org.jetbrains.exposed:exposed-core:_")
     implementation("org.jetbrains.exposed:exposed-core:_")
     implementation("org.jetbrains.exposed:exposed-dao:_")
@@ -44,8 +30,6 @@ dependencies {
     implementation("org.jetbrains.exposed:exposed-jdbc:_")
     implementation("org.jetbrains.kotlin:kotlin-reflect:_")
     implementation("org.jetbrains.kotlinx:kotlinx-datetime:_")
-    implementation("org.jetbrains.kotlinx:kotlinx-serialization-json:_")
-    implementation("org.jetbrains.kotlinx:kotlinx-serialization-json:_")
     implementation("org.kodein.di:kodein-di:_")
     implementation("org.koin:koin-core:_")
     // Keep dependencies sorted to minimize merge conflicts on pull-requests!
@@ -70,7 +54,6 @@ dependencies {
     implementation(Square.retrofit2.converter.moshi)
     implementation(Square.retrofit2.retrofit)
     // Keep dependencies sorted to minimize merge conflicts on pull-requests!
-    kapt(Square.moshi.kotlinCodegen)
 }
 
 tasks.withType<Test> {
@@ -79,8 +62,4 @@ tasks.withType<Test> {
 
 tasks.register("run", JavaExec::class.java) {
     this.main = "playground._mainKt"
-}
-
-apollo {
-    generateKotlinModels.set(true)
 }
