+++ kotlin-jvm/src/main/kotlin/playground/ManualDI.kt
