+++ versions.properties
@@ -5,16 +5,15 @@
 ## Generated by $ ./gradlew refreshVersions
 ## Please, don't put extra comments in that file yet, keeping them is not supported yet.
 
+plugin.com.apollographql.apollo=2.4.1
+
 plugin.com.squareup.sqldelight=1.4.3
 
-
-version.com.beust..klaxon=5.4
-
-plugin.com.apollographql.apollo=2.4.1
+version.com.apollographql.apollo..apollo-coroutines-support=2.4.1
 
 version.com.apollographql.apollo..apollo-runtime=2.4.1
 
-version.com.apollographql.apollo..apollo-coroutines-support=2.4.1
+version.com.beust..klaxon=5.4
 
 version.com.github.ajalt..mordant=1.2.1
 
@@ -39,7 +38,6 @@ version.io.strikt..strikt-core=0.28.0
 version.kotest=4.2.6
 
 version.kotlin=1.4.10
-## # available=1.4.20-M1
 
 version.kotlinpoet=1.6.0
 
@@ -53,12 +51,11 @@ version.kotlinx.html=0.7.2
 
 version.kotlinx.serialization=1.0.0-RC2
 
- version.ktor=1.4.1
+version.ktor=1.4.1
 
 version.moshi=1.11.0
 
 version.okhttp3=4.9.0
-##  # available=4.10.0-RC1
 
 version.org.jetbrains.exposed..exposed-core=0.27.1
 
@@ -75,15 +72,6 @@ version.org.junit.jupiter..junit-jupiter-params=5.7.0
 version.org.kodein.di..kodein-di=7.1.0
 
 version.org.koin..koin-core=2.1.6
-##              # available=2.2.0-alpha-1
-##              # available=2.2.0-beta-1
-##              # available=2.2.0-beta-2
-##              # available=2.2.0-rc-1
-##              # available=2.2.0-rc-2
-##              # available=3.0.0-alpha-1
-##              # available=3.0.0-alpha-2
-##              # available=3.0.0-alpha-3
-##              # available=3.0.0-alpha-4
 
 version.org.mockito..mockito-core=3.5.13
